feat: add paging support to getSellers query (#168)
* feat: add paging support to getSellers query

The `getSellers` query now accepts a `paging` input parameter, allowing the client to specify the range of sellers to retrieve. The response includes a `paging` object with information about the total number of sellers and the range of sellers returned. This change improves the flexibility and efficiency of retrieving seller data.